src/fastapi/requirements.txt
anyio==3.5.0
db-dtypes==1.0.0
pandas-gbq==0.17.4
fastapi==0.75.0
gcsfs==2022.2.0
google_api_python_client==2.47.0
google-cloud==0.34.0
google-cloud-bigquery==3.0.1
matplotlib==3.2.0
nltk==3.6.5
numpy==1.20.3
pandas==1.3.4
Pillow==9.1.0
plotly==5.7.0
protobuf==3.20.1
pydantic==1.9.0
PyJWT==2.3.0
python-decouple==3.6
python-dotenv==0.20.0
requests==2.27.1
setuptools==58.0.4
starlette==0.13.6
streamlit==1.8.1
textblob==0.17.1
tweepy==3.10.0
wordcloud==1.8.1
uvicorn==0.17.5
